# Projeto_2_Biblioteca/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"BasicResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EmprestimoRequestDto": {
				"properties": {
					"livroId": {
						"type": "number",
						"format": "double"
					},
					"usuarioId": {
						"type": "number",
						"format": "double"
					},
					"dataEmprestimo": {
						"type": "string"
					},
					"dataDevolucao": {
						"type": "string"
					}
				},
				"required": [
					"livroId",
					"usuarioId",
					"dataEmprestimo",
					"dataDevolucao"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmprestimoEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"livroId": {
						"type": "number",
						"format": "double"
					},
					"usuarioId": {
						"type": "number",
						"format": "double"
					},
					"dataEmprestimo": {
						"type": "string",
						"format": "date-time"
					},
					"dataDevolucao": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"livroId",
					"usuarioId",
					"dataEmprestimo",
					"dataDevolucao"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoriaRequestDto": {
				"properties": {
					"nome": {
						"type": "string"
					}
				},
				"required": [
					"nome"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoriaEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nome": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"nome"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Documentação",
		"version": "1.0.0",
		"description": "Documentação da API usando Swagger"
	},
	"paths": {
...
		"/emprestimo/Realizar/Emprestimo": {
			"post": {
				"operationId": "RealizarEmprestimo",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmprestimoRequestDto"
							}
						}
					}
				}
			}
		},
		"/emprestimo/Atualizar/Emprestimo": {
			"put": {
				"operationId": "AtualizarEmprestimo",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmprestimoEntity"
							}
						}
					}
				}
			}
		},
		"/emprestimo/Deletar/Emprestimo": {
			"delete": {
				"operationId": "DeleteEmprestimo",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "param",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/emprestimo/filtrar/Emprestimo": {
			"get": {
				"operationId": "FiltrarEmprestimo",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "param",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/emprestimo/Listar/Todos/os/Emprestimos": {
			"get": {
				"operationId": "ListarTodosEmprestimos",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": []
			}
		},
		"/categoria/Cadastrar/Categoria": {
			"post": {
				"operationId": "CadastrarCategoria",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoriaRequestDto"
							}
						}
					}
				}
			}
		},
		"/categoria/Atualizar/Categoria": {
			"put": {
				"operationId": "AtualizarCategoria",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoriaEntity"
							}
						}
					}
				}
			}
		},
		"/categoria/Deletar/Categoria": {
			"delete": {
				"operationId": "DeleteCategoria",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "param",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/categoria/filtrar/Categoria": {
			"get": {
				"operationId": "FiltrarCategoria",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "param",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/categoria/Listar/Todas/as/Categorias": {
			"get": {
				"operationId": "ListarTodasCategorias",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3040/api",
			"description": "Servidor de Desenvolvimento"
		}
	]
}